test: move integration tests to dedicated classes
