+++ scrapers/src/test/java/nl/esciencecenter/rsd/scraper/git/GithubScraperTest.java
@@ -1,5 +1,5 @@
-// SPDX-FileCopyrightText: 2022 - 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
-// SPDX-FileCopyrightText: 2022 - 2023 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2022 - 2024 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
+// SPDX-FileCopyrightText: 2022 - 2024 Netherlands eScience Center
 // SPDX-FileCopyrightText: 2022 Christian Meeßen (GFZ) <christian.meessen@gfz-potsdam.de>
 // SPDX-FileCopyrightText: 2022 Helmholtz Centre Potsdam - GFZ German Research Centre for Geosciences
 //
@@ -8,7 +8,6 @@
 package nl.esciencecenter.rsd.scraper.git;
 
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
@@ -20,47 +19,6 @@ public class GithubScraperTest {
 	private final String repo = "research-software-directory/RSD-as-a-service";
 	private final String repoEmpty = "cmeessen/empty";
 	private final String repoNonEx = "research-software-directory/does-not-exist";
-
-	private final GithubScraper githubScraper = GithubScraper.create(githubUrlPrefix + repo).get();
-	private final GithubScraper githubScraperEmpty = GithubScraper.create(githubUrlPrefix + repoEmpty).get();
-	private final GithubScraper githubScraperNonEx = GithubScraper.create(githubUrlPrefix + repoNonEx).get();
-
-	@Disabled
-	@Test
-	void languages() {
-		final String languages = Assertions.assertDoesNotThrow(() -> githubScraper.languages());
-		Assertions.assertTrue(languages.startsWith("{"));
-		Assertions.assertTrue(languages.endsWith("}"));
-		Assertions.assertTrue(languages.contains("Java"));
-	}
-
-	@Disabled
-	@Test
-	void license() {
-		String license = Assertions.assertDoesNotThrow(() -> githubScraper.basicData().license);
-		Assertions.assertEquals("Apache-2.0", license);
-	}
-
-	@Disabled
-	@Test
-	void contributions() {
-		final CommitsPerWeek contributions = Assertions.assertDoesNotThrow(githubScraper::contributions);
-		// Assertions.assertTrue(contributions.startsWith("[{\"total"));
-	}
-
-	@Disabled
-	@Test
-	void contributionsEmpty() {
-		final CommitsPerWeek contributionsEmpty = Assertions.assertDoesNotThrow(githubScraperEmpty::contributions);
-		// Assertions.assertTrue("[]", contributionsEmpty);
-	}
-
-	@Disabled
-	@Test
-	void contributionsNonEx() {
-		final CommitsPerWeek contributionsNonEx = Assertions.assertDoesNotThrow(githubScraperNonEx::contributions);
-		// Assertions.assertNull(contributionsNonEx);
-	}
 
 	@Test
 	void givenListWithLastPageHeader_whenParsing_thenCorrectPageReturned() {
