+++ scrapers/src/test/java/nl/esciencecenter/rsd/scraper/git/GitlabScraperIT.java
@@ -1,5 +1,5 @@
-// SPDX-FileCopyrightText: 2022 - 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
-// SPDX-FileCopyrightText: 2022 - 2023 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2022 - 2024 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
+// SPDX-FileCopyrightText: 2022 - 2024 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 
-public class GitlabScraperTest {
+public class GitlabScraperIT {
 	private final String baseApiUrl = "https://git.gfz-potsdam.de/api";
 	private final String repo = "swc-bb/swc-templates/swc-l/python";
 	private final GitlabScraper scraper = new GitlabScraper(baseApiUrl, repo);
